Move user_info and logout handlers to login.py
